Refactor internal audio mixer abstraction
